remove pooling

+++ src/napatrackmater/Trackvector.py
@@ -1787,7 +1787,7 @@ def train_mitosis_neural_net(
         y_train_class2,
         y_val_class2,
     ) = train_test_split(
-        features_array.astype(np.float32),
+        features_array,
         labels_array_class1.astype(np.uint8),
         labels_array_class2.astype(np.uint8),
         test_size=0.1,
